@@ -1,46 +1,31 @@
 import { CommonModule } from '@angular/common';
-import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
-import { FormBuilder, FormControl, FormGroup, ReactiveFormsModule } from '@angular/forms';
-import { User } from 'firebase/auth';
+import { Component, OnInit, computed, inject } from '@angular/core';
+import { ReactiveFormsModule } from '@angular/forms';
 
-import { ExtractInitialsPipe } from '../../pipes';
-
-type ProfileEditForm = {
-  displayName: FormControl<string | null>;
-};
+import { RouterModule } from '@angular/router';
+import { ExtractInitialsPipe } from '@expenses-tracker/shared/common';
+import { EditMode, ProfileEditService } from './profile-edit.service';
 
 @Component({
   selector: 'expenses-tracker-profile-edit',
   standalone: true,
-  imports: [CommonModule, ReactiveFormsModule, ExtractInitialsPipe],
-  templateUrl: './profile-edit.component.html',
-  styles: []
+  imports: [CommonModule, ReactiveFormsModule, RouterModule, ExtractInitialsPipe],
+  templateUrl: './profile-edit.component.html'
 })
 export class ProfileEditComponent implements OnInit {
-  formGroup!: FormGroup<ProfileEditForm>;
-  @Input() user!: User;
-  @Output() updateUserInfo$ = new EventEmitter<User>();
-  @Output() cancelEdit$ = new EventEmitter<void>();
-
-  constructor(private _fb: FormBuilder) {}
+  #service = inject(ProfileEditService);
+  editMode = computed(() => this.#service.editMode());
+  user = computed(() => this.#service.user());
 
   ngOnInit() {
-    this.formGroup = this._fb.group<ProfileEditForm>({
-      displayName: this._fb.control<string>(this.user.displayName ?? '', {
-        validators: []
-      })
-    });
+    this.#service.fetchEditMode();
   }
 
-  updateUserInfo() {
-    this.updateUserInfo$.emit({
-      ...(this.formGroup.value as User),
-      uid: this.user.uid
-    } as User);
+  gotoEditProperty(property: EditMode) {
+    this.#service.gotoEditProperty(property);
   }
 
-  cancelEdit() {
-    this.formGroup.reset();
-    this.cancelEdit$.emit();
+  editProfile() {
+    throw new Error('Method not implemented.');
   }
 }
