@@ -1,340 +1,13 @@
 const { createGlobPatternsForDependencies } = require('@nx/angular/tailwind');
 const { join } = require('path');
+const { button, card, colors, form, input, list, overlay } = require('../../libs/shared/common/src/lib/plugins');
 
 /** @type {import('tailwindcss').Config} */
 module.exports = {
-  plugins: [],
+  plugins: [button, card, form, input, list, overlay],
   content: [join(__dirname, 'src/**/!(*.stories|*.spec).{ts,html}'), ...createGlobPatternsForDependencies(__dirname)],
   darkMode: 'class',
   theme: {
-    colors: {
-      'color-page-header-bg': '#0d1117',
-      'color-marketing-icon-primary': '#79c0ff',
-      'color-marketing-icon-secondary': '#1f6feb',
-      'color-diff-blob-addition-num-text': '#e6edf3',
-      'color-diff-blob-addition-fg': '#e6edf3',
-      'color-diff-blob-deletion-num-text': '#e6edf3',
-      'color-diff-blob-deletion-fg': '#e6edf3',
-      'color-diff-blob-expander-icon': '#7d8590',
-      'color-diffstat-addition-bg': '#3fb950',
-      'color-prettylights-syntax-comment': '#8b949e',
-      'color-prettylights-syntax-constant': '#79c0ff',
-      'color-prettylights-syntax-entity': '#d2a8ff',
-      'color-prettylights-syntax-storage-modifier-import': '#c9d1d9',
-      'color-prettylights-syntax-entity-tag': '#7ee787',
-      'color-prettylights-syntax-keyword': '#ff7b72',
-      'color-prettylights-syntax-string': '#a5d6ff',
-      'color-prettylights-syntax-variable': '#ffa657',
-      'color-prettylights-syntax-brackethighlighter-unmatched': '#f85149',
-      'color-prettylights-syntax-invalid-illegal-text': '#f0f6fc',
-      'color-prettylights-syntax-invalid-illegal-bg': '#8e1519',
-      'color-prettylights-syntax-carriage-return-text': '#f0f6fc',
-      'color-prettylights-syntax-carriage-return-bg': '#b62324',
-      'color-prettylights-syntax-string-regexp': '#7ee787',
-      'color-prettylights-syntax-markup-list': '#f2cc60',
-      'color-prettylights-syntax-markup-heading': '#1f6feb',
-      'color-prettylights-syntax-markup-italic': '#c9d1d9',
-      'color-prettylights-syntax-markup-bold': '#c9d1d9',
-      'color-prettylights-syntax-markup-deleted-text': '#ffdcd7',
-      'color-prettylights-syntax-markup-deleted-bg': '#67060c',
-      'color-prettylights-syntax-markup-inserted-text': '#aff5b4',
-      'color-prettylights-syntax-markup-inserted-bg': '#033a16',
-      'color-prettylights-syntax-markup-changed-text': '#ffdfb6',
-      'color-prettylights-syntax-markup-changed-bg': '#5a1e02',
-      'color-prettylights-syntax-markup-ignored-text': '#c9d1d9',
-      'color-prettylights-syntax-markup-ignored-bg': '#1158c7',
-      'color-prettylights-syntax-meta-diff-range': '#d2a8ff',
-      'color-prettylights-syntax-brackethighlighter-angle': '#8b949e',
-      'color-prettylights-syntax-sublimelinter-gutter-mark': '#484f58',
-      'color-prettylights-syntax-constant-other-reference-link': '#a5d6ff',
-      'color-codemirror-text': '#e6edf3',
-      'color-codemirror-bg': '#0d1117',
-      'color-codemirror-gutters-bg': '#0d1117',
-      'color-codemirror-guttermarker-text': '#0d1117',
-      'color-codemirror-guttermarker-subtle-text': '#6e7681',
-      'color-codemirror-linenumber-text': '#7d8590',
-      'color-codemirror-cursor': '#e6edf3',
-      'color-codemirror-matchingbracket-text': '#e6edf3',
-      'color-codemirror-lines-bg': '#0d1117',
-      'color-codemirror-syntax-comment': '#8b949e',
-      'color-codemirror-syntax-constant': '#79c0ff',
-      'color-codemirror-syntax-entity': '#d2a8ff',
-      'color-codemirror-syntax-keyword': '#ff7b72',
-      'color-codemirror-syntax-storage': '#ff7b72',
-      'color-codemirror-syntax-string': '#a5d6ff',
-      'color-codemirror-syntax-support': '#79c0ff',
-      'color-codemirror-syntax-variable': '#ffa657',
-      'color-checks-bg': '#010409',
-      'color-checks-text-primary': '#e6edf3',
-      'color-checks-text-secondary': '#7d8590',
-      'color-checks-text-link': '#2f81f7',
-      'color-checks-btn-icon': '#7d8590',
-      'color-checks-btn-hover-icon': '#e6edf3',
-      'color-checks-input-text': '#7d8590',
-      'color-checks-input-placeholder-text': '#6e7681',
-      'color-checks-input-focus-text': '#e6edf3',
-      'color-checks-input-bg': '#161b22',
-      'color-checks-donut-error': '#f85149',
-      'color-checks-donut-pending': '#d29922',
-      'color-checks-donut-success': '#2ea043',
-      'color-checks-donut-neutral': '#8b949e',
-      'color-checks-dropdown-text': '#e6edf3',
-      'color-checks-dropdown-bg': '#161b22',
-      'color-checks-dropdown-border': '#30363d',
-      'color-checks-dropdown-hover-text': '#e6edf3',
-      'color-checks-dropdown-btn-hover-text': '#e6edf3',
-      'color-checks-header-label-text': '#7d8590',
-      'color-checks-header-label-open-text': '#e6edf3',
-      'color-checks-header-border': '#21262d',
-      'color-checks-header-icon': '#7d8590',
-      'color-checks-line-text': '#7d8590',
-      'color-checks-line-num-text': '#6e7681',
-      'color-checks-line-timestamp-text': '#6e7681',
-      'color-checks-line-selected-num-text': '#2f81f7',
-      'color-checks-line-dt-fm-text': '#ffffff',
-      'color-checks-line-dt-fm-bg': '#9e6a03',
-      'color-checks-gate-text': '#7d8590',
-      'color-checks-gate-waiting-text': '#d29922',
-      'color-checks-step-header-open-bg': '#161b22',
-      'color-checks-step-error-text': '#f85149',
-      'color-checks-step-warning-text': '#d29922',
-      'color-checks-logline-text': '#7d8590',
-      'color-checks-logline-num-text': '#6e7681',
-      'color-checks-logline-debug-text': '#a371f7',
-      'color-checks-logline-error-text': '#7d8590',
-      'color-checks-logline-error-num-text': '#6e7681',
-      'color-checks-logline-warning-text': '#7d8590',
-      'color-checks-logline-warning-num-text': '#d29922',
-      'color-checks-logline-command-text': '#2f81f7',
-      'color-checks-logline-section-text': '#3fb950',
-      'color-checks-ansi-black': '#0d1117',
-      'color-checks-ansi-black-bright': '#161b22',
-      'color-checks-ansi-white': '#b1bac4',
-      'color-checks-ansi-white-bright': '#b1bac4',
-      'color-checks-ansi-gray': '#6e7681',
-      'color-checks-ansi-red': '#ff7b72',
-      'color-checks-ansi-red-bright': '#ffa198',
-      'color-checks-ansi-green': '#3fb950',
-      'color-checks-ansi-green-bright': '#56d364',
-      'color-checks-ansi-yellow': '#d29922',
-      'color-checks-ansi-yellow-bright': '#e3b341',
-      'color-checks-ansi-blue': '#58a6ff',
-      'color-checks-ansi-blue-bright': '#79c0ff',
-      'color-checks-ansi-magenta': '#bc8cff',
-      'color-checks-ansi-magenta-bright': '#d2a8ff',
-      'color-checks-ansi-cyan': '#76e3ea',
-      'color-checks-ansi-cyan-bright': '#b3f0ff',
-      'color-project-header-bg': '#0d1117',
-      'color-project-sidebar-bg': '#161b22',
-      'color-project-gradient-in': '#161b22',
-      'color-mktg-btn-bg': '#f6f8fa',
-      'color-control-border-color-emphasis': '#606771',
-      'color-avatar-stack-fade': '#30363d',
-      'color-avatar-stack-fade-more': '#21262d',
-      'color-select-menu-backdrop-border': '#484f58',
-      'color-select-menu-tap-focus-bg': '#0c2d6b',
-      'color-header-bg': '#161b22',
-      'color-header-divider': '#8b949e',
-      'color-header-logo': '#f0f6fc',
-      'color-header-search-bg': '#0d1117',
-      'color-header-search-border': '#30363d',
-      'color-sidenav-selected-bg': '#21262d',
-      'color-menu-bg-active': '#161b22',
-      'color-timeline-badge-bg': '#21262d',
-      'color-ansi-black': '#484f58',
-      'color-ansi-black-bright': '#6e7681',
-      'color-ansi-white': '#b1bac4',
-      'color-ansi-white-bright': '#ffffff',
-      'color-ansi-gray': '#6e7681',
-      'color-ansi-red': '#ff7b72',
-      'color-ansi-red-bright': '#ffa198',
-      'color-ansi-green': '#3fb950',
-      'color-ansi-green-bright': '#56d364',
-      'color-ansi-yellow': '#d29922',
-      'color-ansi-yellow-bright': '#e3b341',
-      'color-ansi-blue': '#58a6ff',
-      'color-ansi-blue-bright': '#79c0ff',
-      'color-ansi-magenta': '#bc8cff',
-      'color-ansi-magenta-bright': '#d2a8ff',
-      'color-ansi-cyan': '#39c5cf',
-      'color-ansi-cyan-bright': '#56d4dd',
-      'color-btn-text': '#c9d1d9',
-      'color-btn-bg': '#21262d',
-      'color-btn-hover-bg': '#30363d',
-      'color-btn-hover-border': '#8b949e',
-      'color-btn-active-border': '#6e7681',
-      'color-btn-selected-bg': '#161b22',
-      'color-btn-counter-bg': '#30363d',
-      'color-btn-primary-text': '#ffffff',
-      'color-btn-primary-bg': '#238636',
-      'color-btn-primary-hover-bg': '#2ea043',
-      'color-btn-primary-selected-bg': '#238636',
-      'color-btn-primary-icon': '#ffffff',
-      'color-btn-outline-text': '#388bfd',
-      'color-btn-outline-hover-text': '#58a6ff',
-      'color-btn-outline-hover-bg': '#30363d',
-      'color-btn-outline-selected-text': '#ffffff',
-      'color-btn-outline-selected-bg': '#0d419d',
-      'color-btn-outline-disabled-bg': '#0d1117',
-      'color-btn-outline-hover-counter-fg': '#58a6ff',
-      'color-btn-outline-counter-fg': '#388bfd',
-      'color-btn-danger-text': '#f85149',
-      'color-btn-danger-hover-text': '#ffffff',
-      'color-btn-danger-hover-bg': '#da3633',
-      'color-btn-danger-hover-border': '#f85149',
-      'color-btn-danger-hover-icon': '#ffffff',
-      'color-btn-danger-selected-text': '#ffffff',
-      'color-btn-danger-selected-bg': '#b62324',
-      'color-btn-danger-selected-border': '#ff7b72',
-      'color-btn-danger-disabled-bg': '#0d1117',
-      'color-btn-danger-icon': '#f85149',
-      'color-btn-danger-counter-fg': '#f85149',
-      'color-btn-danger-hover-counter-fg': '#ffffff',
-      'color-underlinenav-icon': '#6e7681',
-      'color-action-list-item-danger-hover-text': '#ff7b72',
-      'color-switch-track-disabled-bg': '#21262d',
-      'color-switch-track-fg': '#7d8590',
-      'color-switch-track-disabled-fg': '#010409',
-      'color-switch-track-checked-fg': '#ffffff',
-      'color-switch-track-checked-disabled-fg': '#010409',
-      'color-switch-knob-bg': '#0d1117',
-      'color-switch-knob-border': '#606771',
-      'color-switch-knob-disabled-bg': '#161b22',
-      'color-switch-knob-checked-bg': '#0d1117',
-      'color-switch-knob-checked-disabled-bg': '#161b22',
-      'color-segmented-control-button-bg': '#0d1117',
-      'color-segmented-control-button-hover-bg': '#30363d',
-      'color-segmented-control-button-active-bg': '#21262d',
-      'color-segmented-control-button-selected-border': '#6e7681',
-      'color-tree-view-item-directory-fill': '#7d8590',
-      'color-fg-default': '#e6edf3',
-      'color-fg-muted': '#7d8590',
-      'color-fg-subtle': '#6e7681',
-      'color-fg-on-emphasis': '#ffffff',
-      'color-canvas-default': '#0d1117',
-      'color-canvas-overlay': '#161b22',
-      'color-canvas-inset': '#010409',
-      'color-canvas-subtle': '#161b22',
-      'color-border-default': '#30363d',
-      'color-border-muted': '#21262d',
-      'color-neutral-emphasis-plus': '#6e7681',
-      'color-neutral-emphasis': '#6e7681',
-      'color-accent-fg': '#2f81f7',
-      'color-accent-emphasis': '#1f6feb',
-      'color-success-fg': '#3fb950',
-      'color-success-emphasis': '#238636',
-      'color-attention-fg': '#d29922',
-      'color-attention-emphasis': '#9e6a03',
-      'color-severe-fg': '#db6d28',
-      'color-severe-emphasis': '#bd561d',
-      'color-danger-fg': '#f85149',
-      'color-danger-emphasis': '#da3633',
-      'color-open-fg': '#3fb950',
-      'color-open-emphasis': '#238636',
-      'color-closed-fg': '#f85149',
-      'color-closed-emphasis': '#da3633',
-      'color-done-fg': '#a371f7',
-      'color-done-emphasis': '#8957e5',
-      'color-sponsors-fg': '#db61a2',
-      'color-sponsors-emphasis': '#bf4b8a',
-      'color-primer-fg-disabled': '#484f58',
-      'color-primer-border-active': '#f78166',
-      'color-canvas-default-transparent': 'rgba(13,17,23,0)',
-      'color-diff-blob-addition-num-bg': 'rgba(63,185,80,0.3)',
-      'color-diff-blob-addition-line-bg': 'rgba(46,160,67,0.15)',
-      'color-diff-blob-addition-word-bg': 'rgba(46,160,67,0.4)',
-      'color-diff-blob-deletion-num-bg': 'rgba(248,81,73,0.3)',
-      'color-diff-blob-deletion-line-bg': 'rgba(248,81,73,0.1)',
-      'color-diff-blob-deletion-word-bg': 'rgba(248,81,73,0.4)',
-      'color-diff-blob-hunk-num-bg': 'rgba(56,139,253,0.4)',
-      'color-diffstat-deletion-border': 'rgba(240,246,252,0.1)',
-      'color-diffstat-addition-border': 'rgba(240,246,252,0.1)',
-      'color-search-keyword-hl': 'rgba(210,153,34,0.4)',
-      'color-codemirror-selection-bg': 'rgba(56,139,253,0.4)',
-      'color-codemirror-activeline-bg': 'rgba(110,118,129,0.1)',
-      'color-checks-btn-hover-bg': 'rgba(110,118,129,0.1)',
-      'color-checks-dropdown-shadow': 'rgba(1,4,9,0.3)',
-      'color-checks-dropdown-hover-bg': 'rgba(110,118,129,0.1)',
-      'color-checks-dropdown-btn-hover-bg': 'rgba(110,118,129,0.1)',
-      'color-checks-scrollbar-thumb-bg': 'rgba(110,118,129,0.4)',
-      'color-checks-line-hover-bg': 'rgba(110,118,129,0.1)',
-      'color-checks-line-selected-bg': 'rgba(56,139,253,0.1)',
-      'color-checks-gate-bg': 'rgba(187,128,9,0.15)',
-      'color-checks-logline-error-bg': 'rgba(248,81,73,0.1)',
-      'color-checks-logline-warning-bg': 'rgba(187,128,9,0.15)',
-      'color-project-gradient-out': 'rgba(22,27,34,0)',
-      'color-mktg-btn-shadow-outline': 'rgb(255 255 255 / 25%) 0 0 0 1px inset',
-      'color-mktg-btn-shadow-focus': 'rgb(255 255 255 / 25%) 0 0 0 4px',
-      'color-mktg-btn-shadow-hover-muted': 'rgb(255 255 255) 0 0 0 2px inset',
-      'color-avatar-bg': 'rgba(255,255,255,0.1)',
-      'color-avatar-border': 'rgba(240,246,252,0.1)',
-      'color-topic-tag-border': 'rgba(0,0,0,0)',
-      'color-counter-border': 'rgba(0,0,0,0)',
-      'color-select-menu-tap-highlight': 'rgba(48,54,61,0.5)',
-      'color-header-text': 'rgba(255,255,255,0.7)',
-      'color-input-disabled-bg': 'rgba(110,118,129,0)',
-      'color-btn-border': 'rgba(240,246,252,0.1)',
-      'color-btn-primary-border': 'rgba(240,246,252,0.1)',
-      'color-btn-primary-hover-border': 'rgba(240,246,252,0.1)',
-      'color-btn-primary-disabled-text': 'rgba(255,255,255,0.5)',
-      'color-btn-primary-disabled-bg': 'rgba(35,134,54,0.6)',
-      'color-btn-primary-disabled-border': 'rgba(240,246,252,0.1)',
-      'color-btn-primary-counter-bg': 'rgba(4,38,15,0.2)',
-      'color-btn-outline-hover-border': 'rgba(240,246,252,0.1)',
-      'color-btn-outline-hover-counter-bg': 'rgba(5,29,77,0.2)',
-      'color-btn-outline-selected-border': 'rgba(240,246,252,0.1)',
-      'color-btn-outline-disabled-text': 'rgba(88,166,255,0.5)',
-      'color-btn-outline-disabled-counter-bg': 'rgba(31,111,235,0.05)',
-      'color-btn-outline-counter-bg': 'rgba(5,29,77,0.2)',
-      'color-btn-outline-disabled-counter-fg': 'rgba(47,129,247,0.5)',
-      'color-btn-danger-hover-counter-bg': 'rgba(255,255,255,0.2)',
-      'color-btn-danger-disabled-text': 'rgba(248,81,73,0.5)',
-      'color-btn-danger-disabled-counter-bg': 'rgba(218,54,51,0.05)',
-      'color-btn-danger-counter-bg': 'rgba(73,2,2,0.2)',
-      'color-btn-danger-disabled-counter-fg': 'rgba(248,81,73,0.5)',
-      'color-underlinenav-border-hover': 'rgba(110,118,129,0.4)',
-      'color-action-list-item-inline-divider': 'rgba(48,54,61,0.48)',
-      'color-action-list-item-default-hover-bg': 'rgba(177,186,196,0.12)',
-      'color-action-list-item-default-hover-border': 'rgba(0,0,0,0)',
-      'color-action-list-item-default-active-bg': 'rgba(177,186,196,0.2)',
-      'color-action-list-item-default-active-border': 'rgba(0,0,0,0)',
-      'color-action-list-item-default-selected-bg': 'rgba(177,186,196,0.08)',
-      'color-action-list-item-danger-hover-bg': 'rgba(248,81,73,0.16)',
-      'color-action-list-item-danger-active-bg': 'rgba(248,81,73,0.24)',
-      'color-switch-track-bg': 'rgba(110,118,129,0.1)',
-      'color-switch-track-active-bg': 'rgba(110,118,129,0.4)',
-      'color-switch-track-border': 'rgba(0,0,0,0)',
-      'color-switch-track-checked-bg': 'rgba(31,111,235,0.35)',
-      'color-switch-track-checked-hover-bg': 'rgba(31,111,235,0.5)',
-      'color-switch-track-checked-active-bg': 'rgba(31,111,235,0.65)',
-      'color-switch-track-checked-border': 'rgba(0,0,0,0)',
-      'color-switch-knob-checked-border': 'rgba(31,111,235,0.35)',
-      'color-segmented-control-bg': 'rgba(110,118,129,0.1)',
-      'color-tree-view-item-chevron-hover-bg': 'rgba(177,186,196,0.12)',
-      'color-border-subtle': 'rgba(240,246,252,0.1)',
-      'color-neutral-muted': 'rgba(110,118,129,0.4)',
-      'color-neutral-subtle': 'rgba(110,118,129,0.1)',
-      'color-accent-muted': 'rgba(56,139,253,0.4)',
-      'color-accent-subtle': 'rgba(56,139,253,0.1)',
-      'color-success-muted': 'rgba(46,160,67,0.4)',
-      'color-success-subtle': 'rgba(46,160,67,0.15)',
-      'color-attention-muted': 'rgba(187,128,9,0.4)',
-      'color-attention-subtle': 'rgba(187,128,9,0.15)',
-      'color-severe-muted': 'rgba(219,109,40,0.4)',
-      'color-severe-subtle': 'rgba(219,109,40,0.1)',
-      'color-danger-muted': 'rgba(248,81,73,0.4)',
-      'color-danger-subtle': 'rgba(248,81,73,0.1)',
-      'color-open-muted': 'rgba(46,160,67,0.4)',
-      'color-open-subtle': 'rgba(46,160,67,0.15)',
-      'color-closed-muted': 'rgba(248,81,73,0.4)',
-      'color-closed-subtle': 'rgba(248,81,73,0.15)',
-      'color-done-muted': 'rgba(163,113,247,0.4)',
-      'color-done-subtle': 'rgba(163,113,247,0.1)',
-      'color-sponsors-muted': 'rgba(219,97,162,0.4)',
-      'color-sponsors-subtle': 'rgba(219,97,162,0.1)',
-      'color-primer-canvas-backdrop': 'rgba(1,4,9,0.8)',
-      'color-primer-canvas-sticky': 'rgba(13,17,23,0.95)',
-      'color-primer-border-contrast': 'rgba(255,255,255,0.2)'
-    }
+    colors: { ...colors }
   }
 };
