@@ -1,10 +1,11 @@
 import { Injectable } from '@angular/core';
+import { FirebaseError } from 'firebase/app';
 
 @Injectable({
   providedIn: 'root'
 })
 export class ErrorService {
-  #errorMap = new Map<string, string>([
+  #errorMap = new Map<FirebaseError['code'], string>([
     ['auth/email-already-in-use', 'The email entered is already in use'],
     ['auth/invalid-email', 'Invalid Email address'],
     ['auth/network-request-failed', 'Unable to connect to the service, please try again later'],
