@@ -1,4 +1,4 @@
-import { Injectable, inject } from '@angular/core';
+import { Injectable, computed, inject } from '@angular/core';
 import { FirebaseError } from '@angular/fire/app';
 import { AngularFirestore } from '@angular/fire/compat/firestore';
 import { Timestamp } from '@angular/fire/firestore';
@@ -9,6 +9,7 @@ import { ITransaction } from '@expenses-tracker/shared/interfaces';
 
 import { ContextService } from '../../context/context.service';
 import { ErrorService } from '../../error/error.service';
+import { TransactionMapper } from '../firestore.utils';
 
 @Injectable({
   providedIn: 'root'
@@ -17,54 +18,30 @@ export class FirestoreTransactionService {
   #firestore = inject(AngularFirestore);
   #context = inject(ContextService);
   #error = inject(ErrorService);
+  user = computed(() => this.#context.user());
+  pocketbook = computed(() => this.#context.pocketbook());
 
   watchTransactionList$() {
-    return this.#context.watchPocketbook$().pipe(
-      switchMap(pocketbook =>
-        !pocketbook
-          ? of([])
-          : this.#firestore
-              .collection<Partial<ITransaction<Timestamp>>>(Collections.Transaction, ref =>
-                ref
-                  .where('pocketbookId', '==', pocketbook?.id ?? '')
-                  .orderBy('timestamp', 'desc')
-              )
-              .valueChanges()
-              .pipe(
-                map(transactionList =>
-                  transactionList.map(
-                    t =>
-                      ({
-                        ...t,
-                        timestamp: t.timestamp?.toDate()
-                      } as ITransaction)
-                  )
-                ),
-                catchError(({ code }: FirebaseError) =>
-                  throwError(() => new Error(this.#error.getError(code)))
-                )
-              )
-      )
-    );
+    return !this.pocketbook()
+      ? of([])
+      : this.#firestore
+          .collection<Partial<ITransaction<Timestamp>>>(Collections.Transaction, ref =>
+            ref.where('pocketbookId', '==', this.pocketbook()?.id ?? '').orderBy('timestamp', 'desc')
+          )
+          .valueChanges()
+          .pipe(
+            map(transactionList => transactionList.map(txn => TransactionMapper(txn as ITransaction<Timestamp>))),
+            catchError(({ code }: FirebaseError) => throwError(() => new Error(this.#error.getError(code))))
+          );
   }
 
   watchTransaction$(transactionId: string) {
     return this.#firestore
-      .collection<ITransaction<Timestamp>>(Collections.Transaction, ref =>
-        ref.where('id', '==', transactionId ?? '')
-      )
+      .collection<ITransaction<Timestamp>>(Collections.Transaction, ref => ref.where('id', '==', transactionId ?? ''))
       .valueChanges()
       .pipe(
-        map(
-          ([txn]) =>
-            ({
-              ...txn,
-              timestamp: (txn?.timestamp as Timestamp)?.toDate()
-            } as ITransaction)
-        ),
-        catchError(({ code }: FirebaseError) =>
-          throwError(() => new Error(this.#error.getError(code)))
-        )
+        map(([txn]) => TransactionMapper(txn as ITransaction<Timestamp>)),
+        catchError(({ code }: FirebaseError) => throwError(() => new Error(this.#error.getError(code))))
       );
   }
 
@@ -74,37 +51,39 @@ export class FirestoreTransactionService {
     direction,
     message,
     paymentMode,
-    timestamp
+    timestamp = new Date()
   }: Partial<ITransaction>) {
     const docId = this.#firestore.createId();
-    return this.#context.watchPocketbook$().pipe(
-      switchMap(pocketbook =>
-        this.#firestore
-          .collection<Partial<ITransaction>>(Collections.Transaction)
-          .doc(docId)
-          .set({
-            id: docId,
-            timestamp,
-            amount: amount ?? 0,
-            category: category ?? '',
-            direction: direction ?? 'expense',
-            paymentMode: paymentMode ?? 'card',
-            message: message ?? '',
-            pocketbookId: pocketbook?.id ?? ''
-          })
-      ),
-      map(
-        () =>
-          ({
-            amount,
-            category,
-            direction,
-            id: docId,
-            message,
-            timestamp
-          } as ITransaction)
-      )
-    );
+
+    return !this.pocketbook()
+      ? throwError(() => new Error('No pocketbook set in context'))
+      : from(
+          this.#firestore
+            .collection<Partial<ITransaction>>(Collections.Transaction)
+            .doc(docId)
+            .set({
+              id: docId,
+              timestamp,
+              amount: amount ?? 0,
+              category: category ?? '',
+              direction: direction ?? 'expense',
+              message: message ?? '',
+              pocketbookId: this.pocketbook()?.id ?? ''
+            })
+        ).pipe(
+          map(
+            () =>
+              ({
+                amount,
+                category,
+                direction,
+                id: docId,
+                message,
+                timestamp,
+                pocketbookId: this.pocketbook()?.id
+              }) as ITransaction
+          )
+        );
   }
 
   updateTransaction$(transaction: Partial<ITransaction>) {
@@ -124,10 +103,7 @@ export class FirestoreTransactionService {
 
   deleteTransaction$(transactionId: string) {
     return from(
-      this.#firestore
-        .collection<Partial<ITransaction>>(Collections.Transaction)
-        .doc(transactionId)
-        .delete()
+      this.#firestore.collection<Partial<ITransaction>>(Collections.Transaction).doc(transactionId).delete()
     ).pipe(
       catchError(({ code }: FirebaseError) => {
         console.error({ code });
