@@ -1,61 +1,62 @@
+import { Dialog, DialogModule } from '@angular/cdk/dialog';
 import { CommonModule } from '@angular/common';
-import { Component, Input, OnInit, inject } from '@angular/core';
+import { Component, Input, OnDestroy, inject, signal } from '@angular/core';
 import { User } from 'firebase/auth';
-import { BehaviorSubject, EMPTY, Observable, of } from 'rxjs';
+import { Subscription } from 'rxjs';
 
-import { ExtractInitialsPipe } from '@expenses-tracker/features/profile';
+import { ExtractInitialsPipe } from '@expenses-tracker/shared/common';
+import { TooltipModule } from '@expenses-tracker/shared/common';
 import { IPocketbook } from '@expenses-tracker/shared/interfaces';
 
+import { PocketbookContributorsDialogComponent } from './pocketbook-contributors-dialog/pocketbook-contributors-dialog.component';
 import { PocketbookListItemService } from './pocketbook-list-item.service';
 
 @Component({
   selector: 'expenses-tracker-pocketbook-list-item',
   standalone: true,
-  imports: [CommonModule, ExtractInitialsPipe],
+  imports: [CommonModule, ExtractInitialsPipe, TooltipModule, DialogModule],
   providers: [PocketbookListItemService],
   templateUrl: './pocketbook-list-item.component.html'
 })
-export class PocketbookListItemComponent implements OnInit {
-  #pocketbook$ = new BehaviorSubject<IPocketbook | null>(null);
-  #isOwner$ = new BehaviorSubject<boolean>(false);
-  collaboratorList$!: Observable<User[]>;
-  owner$!: Observable<User | null>;
+export class PocketbookListItemComponent implements OnDestroy {
+  pocketbook = signal<IPocketbook | null>(null);
+  isOwner = signal(false);
+  collaboratorList = signal<User[]>([]);
+  owner = signal<User | null>(null);
+  #deleteSubscription$!: Subscription;
+  #deleteDialogSubscription$!: Subscription;
+  #contributorsDialogSubscription$!: Subscription;
 
-  @Input() set pocketbook(value: IPocketbook | null) {
-    this.#pocketbook$.next(value);
-  }
-  get pocketbook() {
-    return this.#pocketbook$.getValue();
+  @Input() set pocketbookInput(value: IPocketbook | null) {
+    this.pocketbook.set(value);
   }
 
-  @Input() set isOwner(value: boolean) {
-    this.#isOwner$.next(value);
-  }
-  get isOwner() {
-    return this.#isOwner$.getValue();
+  @Input() set isOwnerInput(value: boolean) {
+    this.isOwner.set(value);
   }
 
   #service = inject(PocketbookListItemService);
-
-  ngOnInit() {
-    this.#service.initializeComponent(this.pocketbook);
-    this.collaboratorList$ = this.#service.watchCollaboratorList$();
-    this.owner$ = this.#service.watchOwner$();
-  }
+  #dialog = inject(Dialog);
 
   gotoEditPocketbook() {
-    this.#service.gotoEditPocketbook(this.pocketbook as IPocketbook);
+    this.#service.gotoEditPocketbook(this.pocketbook() as IPocketbook);
   }
 
-  deletePocketbook() {
-    const result = confirm('Delete Pocketbook?');
-    const deleteStream = result
-      ? this.#service.deletePocketbook$((this.pocketbook as IPocketbook)?.id ?? '')
-      : EMPTY;
-    deleteStream.subscribe();
+  showContributors() {
+    const dialogRef = this.#dialog.open(PocketbookContributorsDialogComponent, {
+      data: this.pocketbook(),
+      disableClose: true
+    });
+    this.#contributorsDialogSubscription$ = dialogRef.closed.subscribe();
   }
 
   gotoPocketbookDetail() {
-    this.#service.gotoPocketbook(this.pocketbook?.id ?? '');
+    this.#service.gotoPocketbook(this.pocketbook() as IPocketbook);
+  }
+
+  ngOnDestroy() {
+    this.#deleteSubscription$?.unsubscribe();
+    this.#deleteDialogSubscription$?.unsubscribe();
+    this.#contributorsDialogSubscription$?.unsubscribe();
   }
 }
