@@ -1,15 +1,7 @@
+import { Location } from '@angular/common';
 import { Injectable, inject } from '@angular/core';
 import { Router } from '@angular/router';
-import {
-  BehaviorSubject,
-  Observable,
-  catchError,
-  distinctUntilChanged,
-  map,
-  of,
-  switchMap,
-  tap
-} from 'rxjs';
+import { BehaviorSubject, Observable, map, of, switchMap } from 'rxjs';
 
 import { ContextService, FirestoreService } from '@expenses-tracker/core';
 import { NotificationService, RoutePaths } from '@expenses-tracker/shared/common';
@@ -32,6 +24,7 @@ export class TransactionEditService {
   #firestore = inject(FirestoreService);
   #notification = inject(NotificationService);
   #router = inject(Router);
+  #location = inject(Location);
 
   #setFlags(flags: ComponentFlags) {
     this.#flags = flags;
@@ -47,14 +40,13 @@ export class TransactionEditService {
   }
 
   watchTransaction$() {
-    const transactionId = this.#router.url.match(/pocketbook\/(\w+)\//)?.at(1);
+    const transactionId = this.#location
+      .path()
+      .match(/pocketbook\/(\w+)\//)
+      ?.at(1);
     return this.#context
       .watchTransaction$()
-      .pipe(
-        switchMap(txn =>
-          !txn ? this.#firestore.watchTransaction$(transactionId ?? '') : of(txn)
-        )
-      );
+      .pipe(switchMap(txn => (!txn ? this.#firestore.watchTransaction$(transactionId ?? '') : of(txn))));
   }
 
   patchValues$(): Observable<TransactionFormFields> {
@@ -81,59 +73,57 @@ export class TransactionEditService {
     });
 
     const t = this.#context.getTransaction();
-    return this.#firestore.updateTransaction$({ ...t, ...transaction }).pipe(
-      switchMap(() =>
-        this.#context.watchPocketbook$().pipe(
-          distinctUntilChanged(
-            previous => !(previous?.transactionList ?? []).includes(transaction.id)
-          ),
-          switchMap(pocketbook =>
-            this.#firestore.updatePocketbook$({
-              ...pocketbook,
-              balance: this.#context.updateTransactionCalculateBalance({
-                old: t as ITransaction,
-                new: transaction
-              })
-            })
-          )
-        )
-      ),
-      tap(() => {
-        this.#notification.success({
-          title: 'Transaction edited',
-          description: 'The transaction has been updated successfully'
-        });
-        this.resetFlags();
-        this.#router.navigate([
-          RoutePaths.Pocketbook,
-          this.#context.getPocketbook()?.id ?? 'null',
-          RoutePaths.Transaction,
-          RoutePaths.EntityList
-        ]);
-      }),
-      catchError(error => {
-        this.#notification.error({
-          description: `${error}.`,
-          title: 'Error adding transaction'
-        });
-        this.#setFlags({
-          ...this.#flags,
-          editTransaction: {
-            ...this.#flags.editTransaction,
-            loading: false,
-            fail: true,
-            visible: true
-          }
-        });
-        return of(error);
-      })
-    );
+    // return this.#firestore.updateTransaction$({ ...t, ...transaction }).pipe(
+    //   switchMap(() =>
+    //     this.#context.watchPocketbook$().pipe(
+    //       distinctUntilChanged(previous => !(previous?.transactionList ?? []).includes(transaction.id)),
+    //       switchMap(pocketbook =>
+    //         this.#firestore.updatePocketbook$({
+    //           ...pocketbook,
+    //           balance: this.#context.updateTransactionCalculateBalance({
+    //             old: t as ITransaction,
+    //             new: transaction
+    //           })
+    //         })
+    //       )
+    //     )
+    //   ),
+    //   tap(() => {
+    //     this.#notification.success({
+    //       title: 'Transaction edited',
+    //       description: 'The transaction has been updated successfully'
+    //     });
+    //     this.resetFlags();
+    //     this.#router.navigate([
+    //       RoutePaths.Pocketbook,
+    //       this.#context.getPocketbook()?.id ?? 'null',
+    //       RoutePaths.Transaction,
+    //       RoutePaths.EntityList
+    //     ]);
+    //   }),
+    //   catchError(error => {
+    //     this.#notification.error({
+    //       description: `${error}.`,
+    //       title: 'Error adding transaction'
+    //     });
+    //     this.#setFlags({
+    //       ...this.#flags,
+    //       editTransaction: {
+    //         ...this.#flags.editTransaction,
+    //         loading: false,
+    //         fail: true,
+    //         visible: true
+    //       }
+    //     });
+    //     return of(error);
+    //   })
+    // );
   }
 
   cancelEditTransaction() {
     this.#router.navigate([
       RoutePaths.Pocketbook,
-      this.#context.getPocketbook()?.id,
+      // this.#context.getPocketbook()?.id,
       RoutePaths.Transaction,
       RoutePaths.EntityList
     ]);
