@@ -1,6 +1,5 @@
 import { CommonModule } from '@angular/common';
 import { Component } from '@angular/core';
-import { MatDialogModule } from '@angular/material/dialog';
 
 /**
  * @deprecated - Removed with material
@@ -8,7 +7,7 @@ import { MatDialogModule } from '@angular/material/dialog';
 @Component({
   selector: 'expenses-tracker-transaction-delete-dialog',
   standalone: true,
-  imports: [CommonModule, MatDialogModule],
+  imports: [CommonModule],
   template: `<div
     class="grid p-6 gap-6 rounded-lg border bg-et-layer-alternate text-et-color-fg-default border-et-color-button-default-border">
     <div class="flex items-center gap-2">
@@ -20,7 +19,6 @@ import { MatDialogModule } from '@angular/material/dialog';
     <div class="flex flex-col">
       <div class="flex items-center justify-center gap-6">
         <button
-          [mat-dialog-close]="true"
           type="button"
           class="flex justify-center items-center w-full gap-2 px-6 py-3 text-center text-et-color-button-danger bg-et-color-button-danger-bg rounded-lg cursor-pointer font-bold">
           <span class="material-icons min-w-[24px]">delete</span>
