@@ -19,7 +19,7 @@ export class TransactionListItemService {
     this.#context.setTransaction(transaction);
     this.#router.navigate([
       RoutePaths.Pocketbook,
-      this.#context.getPocketbook()?.id,
+      // this.#context.getPocketbook()?.id,
       RoutePaths.Transaction,
       this.#context.getTransaction()?.id,
       RoutePaths.EntityEdit
@@ -27,41 +27,41 @@ export class TransactionListItemService {
   }
 
   deleteTransaction$(transaction: ITransaction) {
-    return this.#firestore.deleteTransaction$(transaction?.id ?? '').pipe(
-      switchMap(() =>
-        this.#context.watchPocketbook$().pipe(
-          distinctUntilChanged(previous =>
-            (previous?.transactionList ?? []).includes(transaction?.id ?? '')
-          ),
-          switchMap(pocketbook =>
-            this.#firestore.updatePocketbook$({
-              ...pocketbook,
-              transactionList: pocketbook?.transactionList.filter(t => t !== transaction?.id),
-              balance: this.#context.deleteTransactionCalculateBalance(transaction)
-            })
-          )
-        )
-      ),
-      tap(() => {
-        this.#notification.success({
-          title: 'Deleted Successfully',
-          description: 'Transaction successfully deleted'
-        });
-        this.#context.setTransaction(null);
-        this.#router.navigate([
-          RoutePaths.Pocketbook,
-          this.#context.getPocketbook()?.id,
-          RoutePaths.Transaction,
-          RoutePaths.EntityList
-        ]);
-      }),
-      catchError(error => {
-        this.#notification.error({
-          title: 'Delete Failed',
-          description: 'Unable to delete the transaction'
-        });
-        return throwError(() => new Error(error));
-      })
-    );
+    //   return this.#firestore.deleteTransaction$(transaction?.id ?? '').pipe(
+    //     switchMap(() =>
+    //       this.#context.watchPocketbook$().pipe(
+    //         distinctUntilChanged(previous =>
+    //           (previous?.transactionList ?? []).includes(transaction?.id ?? '')
+    //         ),
+    //         switchMap(pocketbook =>
+    //           this.#firestore.updatePocketbook$({
+    //             ...pocketbook,
+    //             transactionList: pocketbook?.transactionList.filter(t => t !== transaction?.id),
+    //             balance: this.#context.deleteTransactionCalculateBalance(transaction)
+    //           })
+    //         )
+    //       )
+    //     ),
+    //     tap(() => {
+    //       this.#notification.success({
+    //         title: 'Deleted Successfully',
+    //         description: 'Transaction successfully deleted'
+    //       });
+    //       this.#context.setTransaction(null);
+    //       this.#router.navigate([
+    //         RoutePaths.Pocketbook,
+    //         this.#context.getPocketbook()?.id,
+    //         RoutePaths.Transaction,
+    //         RoutePaths.EntityList
+    //       ]);
+    //     }),
+    //     catchError(error => {
+    //       this.#notification.error({
+    //         title: 'Delete Failed',
+    //         description: 'Unable to delete the transaction'
+    //       });
+    //       return throwError(() => new Error(error));
+    //     })
+    //   );
   }
 }
