@@ -1,4 +1,4 @@
-import { Injectable, inject } from '@angular/core';
+import { Injectable, computed, inject } from '@angular/core';
 import { Router } from '@angular/router';
 
 import { ContextService } from '@expenses-tracker/core';
@@ -10,10 +10,7 @@ import { RoutePaths } from '@expenses-tracker/shared/common';
 export class NavbarService {
   #context = inject(ContextService);
   #router = inject(Router);
-
-  getUser$() {
-    return this.#context.watchUser$();
-  }
+  user = computed(() => this.#context.user());
 
   gotoHome() {
     this.#router.navigate([RoutePaths.Home]);
