@@ -1,5 +1,5 @@
 export * from './lib/common';
-
+export * from './lib/forms';
 export * from './lib/flags';
 export * from './lib/pocketbook';
 export * from './lib/transaction';
